Cleanup ui code
